requirements.txt
absl-py==0.7.1
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
dask==2.17.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
filelock==3.0.12
easydict==1.9
entrypoints==0.3
gast==0.2.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.13.2
grpcio==1.25.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
llvmlite==0.32.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numba==0.49.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python==3.4.4.19
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.0
qtconsole==4.5.2
regex==2019.11.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
sacremoses==0.0.43
scikit-image==0.14.2
scikit-learn==0.23.0
scipy=1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
tensorboardX==1.9
tensorflow-estimator==2.0.1
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
toolz==0.9.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
transformers==4.3.3
umap_learn==0.4.2
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
